Refactor creation of render pipeline
* find and check bind groups for shaders
* create and set render pipeline
* set shaders in render pipeline instead of keeping in variable
