+++ src/renderer/src/pipeline/pipeline_layout.rs
@@ -14,10 +14,20 @@ impl PipelineLayout {
     pub fn from_shaders(shaders: Vec<&Shader>) -> Result<Self, ShaderError> {
         let mut bind_groups = HashMap::<u32, BindGroupDescriptor>::new();
         for shader in shaders {
-            let shader_layout = shader.layout();
-            for _bind_group in shader_layout.bind_groups.iter() {
-                // bind_groups
-                panic!("Implement mapping of bind groups");
+            let mut shader_layout = shader.layout();
+            for shader_bind_group in shader_layout.bind_groups.iter_mut() {
+                match bind_groups.get_mut(&shader_bind_group.index) {
+                    Some(bind_group) => {
+                        for shader_binding in shader_bind_group.bindings.iter() {
+                            if !shader_bind_group.contains(shader_binding) {
+                                bind_group.bindings.push(shader_binding.clone());
+                            }
+                        }
+                    }
+                    None => {
+                        bind_groups.insert(shader_bind_group.index, shader_bind_group.clone());
+                    }
+                }
             }
         }
