+++ src/renderer/src/renderer.rs
@@ -1,8 +1,8 @@
-use std::{sync::Arc};
+use std::{sync::Arc, collections::HashMap};
 
 use wgpu::{BufferUsage, util::DeviceExt};
 
-use crate::{PassDescriptor, RenderPass, Shader, ShaderStage, Surface, Texture, TextureDescriptor, TextureFormat, VertexBuffer};
+use crate::{BindGroupDescriptor, PassDescriptor, PipelineDescriptor, RenderPass, Shader, ShaderStage, Surface, Texture, TextureDescriptor, TextureFormat, VertexBuffer};
 
 pub struct RenderPassHandle {}
 
@@ -86,8 +86,78 @@ impl Renderer {
         }
     }
 
-    pub fn create_render_pipeline(&mut self) {
 
+    /// For now this returns the internal wgpu RenderPipeline until it's a bit clearer
+    /// where and how to handle them
+    pub fn create_pipeline(
+        &mut self,
+        pipeline_descriptor: &PipelineDescriptor,
+    ) -> wgpu::RenderPipeline {
+        let layout = pipeline_descriptor.get_layout().unwrap();
+        let bind_group_layouts = layout.bind_groups
+            .iter()
+            .map(|bind_group| create_bind_group_layout(&self.device, bind_group))
+            .collect::<Vec<wgpu::BindGroupLayout>>();
+
+        let bind_group_layouts_ref = bind_group_layouts
+            .iter()
+            .map(|layout| layout)
+            .collect::<Vec<&wgpu::BindGroupLayout>>();
+
+        let pipeline_layout =
+            self.device
+                .create_pipeline_layout(&wgpu::PipelineLayoutDescriptor {
+                    label: None,
+                    bind_group_layouts: bind_group_layouts_ref.as_slice(),
+                    push_constant_ranges: &[],
+                });
+
+        // set up shaders
+        let vertex_stage = wgpu::ProgrammableStageDescriptor {
+            module: &pipeline_descriptor.vertex_shader.module,
+            entry_point: "main",
+        };
+
+        let vertex_state = wgpu::VertexStateDescriptor {
+            index_format: pipeline_descriptor.index_format.into(),
+            vertex_buffers: &[],
+        };
+
+        let fragment_stage = wgpu::ProgrammableStageDescriptor {
+            module: &pipeline_descriptor.fragment_shader.module,
+            entry_point: "main",
+        };
+
+        let rasterization_state = pipeline_descriptor.rasterization_state
+            .as_ref()
+            .map(|desc| desc.into());
+
+        let color_states = pipeline_descriptor.color_states
+            .iter()
+            .map(|c| c.into())
+            .collect::<Vec<wgpu::ColorStateDescriptor>>();
+
+        let depth_stencil_state = pipeline_descriptor
+            .depth_stencil_state
+            .as_ref()
+            .map(|desc| desc.into());
+
+        let render_pipeline_descriptor = wgpu::RenderPipelineDescriptor {
+            label: pipeline_descriptor.label.as_ref().map(|label| label.as_str()),
+            layout: Some(&pipeline_layout),
+            vertex_stage,
+            fragment_stage: Some(fragment_stage),
+            rasterization_state,
+            primitive_topology: pipeline_descriptor.primitive_topology.into(),
+            color_states: &color_states,
+            depth_stencil_state,
+            vertex_state,
+            sample_count: pipeline_descriptor.sample_count,
+            sample_mask: pipeline_descriptor.sample_mask,
+            alpha_to_coverage_enabled: pipeline_descriptor.alpha_to_coverage_enabled,
+        };
+
+        self.device.create_render_pipeline(&render_pipeline_descriptor)
     }
 
     /// Creates a new vertex buffer
@@ -163,6 +233,30 @@ impl Renderer {
         let descriptor = TextureDescriptor::depth(width, height);
         Texture::new(descriptor, &self.device)
     }
+}
+
+fn create_bind_group_layout(
+    device: &wgpu::Device,
+    descriptor: &BindGroupDescriptor,
+) -> wgpu::BindGroupLayout {
+    let entries = descriptor.bindings
+        .iter()
+        .map(|binding| {
+            wgpu::BindGroupLayoutEntry {
+                binding: binding.index,
+                visibility: binding.shader_stage.into(),
+                ty: (&binding.binding_type).into(),
+                count: None,
+            }
+        })
+        .collect::<Vec<wgpu::BindGroupLayoutEntry>>();
+
+    let bind_group_layout_descriptor = wgpu::BindGroupLayoutDescriptor {
+        entries: entries.as_slice(),
+        label: None,
+    };
+
+    device.create_bind_group_layout(&bind_group_layout_descriptor)
 }
 
 /// Creates a new internal RenderPass reference
